+++ apps/frontend/src/app/interviewer/page.tsx
@@ -1,4 +1,4 @@
-import InterviewerCard from "./_components/InterviewerCard";
+import InterviewerCard from "@/_components/interviewer/InterviewerCard";
 import styles from "./page.module.css";
 import Text from "@repo/ui/Text";
